create tests; rename files; separate server from tests
